+++ packages/sdks/react-sdk/src/components/UserProfile.tsx
@@ -25,7 +25,18 @@ const UserProfileWC = lazy(async () => {
 
 const UserProfile = React.forwardRef<HTMLElement, UserProfileProps>(
   (
-    { logger, theme, locale, debug, widgetId, onLogout, styleId, onReady },
+    {
+      logger,
+      theme,
+      locale,
+      debug,
+      widgetId,
+      onLogout,
+      styleId,
+      onReady,
+      client,
+      form,
+    },
     ref,
   ) => {
     const [innerRef, setInnerRef] = useState(null);
@@ -90,6 +101,8 @@ const UserProfile = React.forwardRef<HTMLElement, UserProfileProps>(
             'debug.attr': debug,
             'styleId.attr': styleId,
             'refreshCookieName.attr': refreshCookieName,
+            'client.attr': client,
+            'form.attr': form,
             // props
             'logger.prop': logger,
           }}