+++ packages/widgets/user-profile-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initGenericFlowButtonMixin.ts
@@ -4,11 +4,15 @@ import {
   ModalDriver,
 } from '@descope/sdk-component-drivers';
 import { compose, createSingletonMixin } from '@descope/sdk-helpers';
-import { localeMixin, loggerMixin, modalMixin } from '@descope/sdk-mixins';
+import {
+  localeMixin,
+  loggerMixin,
+  modalMixin,
+  flowInputMixin,
+} from '@descope/sdk-mixins';
 import { flowSyncThemeMixin } from '../../flowSyncThemeMixin';
 import { initWidgetRootMixin } from './initWidgetRootMixin';
 import { stateManagementMixin } from '../../stateManagementMixin';
-import { createFlowTemplate } from '../../helpers';
 import { getUserId } from '../../../state/selectors';
 
 export const initGenericFlowButtonMixin = createSingletonMixin(
@@ -20,6 +24,7 @@ export const initGenericFlowButtonMixin = createSingletonMixin(
       modalMixin,
       loggerMixin,
       initWidgetRootMixin,
+      flowInputMixin,
     )(superclass) {
       #modal: ModalDriver;
 
@@ -49,7 +54,10 @@ export const initGenericFlowButtonMixin = createSingletonMixin(
       }
 
       #initModal() {
-        this.#modal = this.createModal({ 'data-id': 'generic-flow-modal' });
+        this.#modal = this.createModal({
+          'data-id': 'generic-flow-modal',
+          'close-on-outside-click': 'true',
+        });
         this.#modal.afterClose = () => {
           this.#removePageUpdatedCallback?.();
           this.#removePageUpdatedCallback = null;
@@ -68,18 +76,7 @@ export const initGenericFlowButtonMixin = createSingletonMixin(
 
       #initModalContent(flowId: string, userId: string) {
         this.#modal.setContent(
-          createFlowTemplate({
-            locale: this.locale,
-            projectId: this.projectId,
-            flowId,
-            baseUrl: this.baseUrl,
-            baseStaticUrl: this.baseStaticUrl,
-            baseCdnUrl: this.baseCdnUrl,
-            refreshCookieName: this.refreshCookieName,
-            theme: this.theme,
-            'style-id': this.styleId,
-            client: { userId },
-          }),
+          this.createFlowTemplate({ flowId, client: { userId } }),
         );
         const cb = () => this.#onModalNeeded();
         this.#removePageUpdatedCallback = this.#flow.onPageUpdated(cb);