+++ packages/widgets/user-profile-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initPhoneUserAttrMixin.ts
@@ -13,9 +13,9 @@ import {
   loggerMixin,
   modalMixin,
   cookieConfigMixin,
+  flowInputMixin,
 } from '@descope/sdk-mixins';
 import { getPhone, getPhoneBadgeLabel } from '../../../state/selectors';
-import { createFlowTemplate } from '../../helpers';
 import { stateManagementMixin } from '../../stateManagementMixin';
 import { initWidgetRootMixin } from './initWidgetRootMixin';
 import { flowSyncThemeMixin } from '../../flowSyncThemeMixin';
@@ -30,6 +30,7 @@ export const initPhoneUserAttrMixin = createSingletonMixin(
       initWidgetRootMixin,
       cookieConfigMixin,
       modalMixin,
+      flowInputMixin,
     )(superclass) {
       phoneUserAttr: UserAttributeDriver;
 
@@ -44,7 +45,10 @@ export const initPhoneUserAttrMixin = createSingletonMixin(
       #initEditModal() {
         if (!this.phoneUserAttr.editFlowId) return;
 
-        this.#editModal = this.createModal({ 'data-id': 'edit-phone' });
+        this.#editModal = this.createModal({
+          'data-id': 'edit-phone',
+          'close-on-outside-click': 'true',
+        });
         this.#editFlow = new FlowDriver(
           () => this.#editModal.ele?.querySelector('descope-wc'),
           { logger: this.logger },
@@ -56,17 +60,7 @@ export const initPhoneUserAttrMixin = createSingletonMixin(
 
       #initEditModalContent() {
         this.#editModal.setContent(
-          createFlowTemplate({
-            locale: this.locale,
-            projectId: this.projectId,
-            flowId: this.phoneUserAttr.editFlowId,
-            baseUrl: this.baseUrl,
-            baseStaticUrl: this.baseStaticUrl,
-            baseCdnUrl: this.baseCdnUrl,
-            refreshCookieName: this.refreshCookieName,
-            theme: this.theme,
-            'style-id': this.styleId,
-          }),
+          this.createFlowTemplate({ flowId: this.phoneUserAttr.editFlowId }),
         );
         this.#editFlow.onSuccess(() => {
           this.#editModal.close();
@@ -77,7 +71,10 @@ export const initPhoneUserAttrMixin = createSingletonMixin(
       #initDeleteModal() {
         if (!this.phoneUserAttr.deleteFlowId) return;
 
-        this.#deleteModal = this.createModal({ 'data-id': 'delete-phone' });
+        this.#deleteModal = this.createModal({
+          'data-id': 'delete-phone',
+          'close-on-outside-click': 'true',
+        });
         this.#deleteFlow = new FlowDriver(
           () => this.#deleteModal.ele?.querySelector('descope-wc'),
           { logger: this.logger },
@@ -89,17 +86,7 @@ export const initPhoneUserAttrMixin = createSingletonMixin(
 
       #initDeleteModalContent() {
         this.#deleteModal.setContent(
-          createFlowTemplate({
-            locale: this.locale,
-            projectId: this.projectId,
-            flowId: this.phoneUserAttr.deleteFlowId,
-            baseUrl: this.baseUrl,
-            baseStaticUrl: this.baseStaticUrl,
-            baseCdnUrl: this.baseCdnUrl,
-            refreshCookieName: this.refreshCookieName,
-            theme: this.theme,
-            'style-id': this.styleId,
-          }),
+          this.createFlowTemplate({ flowId: this.phoneUserAttr.deleteFlowId }),
         );
         this.#deleteFlow.onSuccess(() => {
           this.#deleteModal.close();