+++ packages/sdks/web-component/CHANGELOG.md
@@ -2,6 +2,57 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.70.0](https://github.com/descope/descope-js/compare/web-component-3.69.5...web-component-3.70.0) (2026-07-05)
+
+
+### Features
+
+* **web-component:** warn on multiple same-flow components on a page RELEASE ([#1431](https://github.com/descope/descope-js/issues/1431)) ([2f91a28](https://github.com/descope/descope-js/commit/2f91a2800e927589328e98c25219fafb2d63c4b2)), closes [#14040](https://github.com/descope/descope-js/issues/14040) [descope/etc#14040](https://github.com/descope/etc/issues/14040)
+
+## [3.69.5](https://github.com/descope/descope-js/compare/web-component-3.69.4...web-component-3.69.5) (2026-06-29)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.23.0`
+## [3.69.4](https://github.com/descope/descope-js/compare/web-component-3.69.3...web-component-3.69.4) (2026-06-28)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.9.0`
+* `sdk-mixins` updated to version `0.22.0`
+* `web-js-sdk` updated to version `1.51.1`
+## [3.69.3](https://github.com/descope/descope-js/compare/web-component-3.69.2...web-component-3.69.3) (2026-06-25)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.51.0`
+
+### Bug Fixes
+
+* **web-component:** clear loading state after flow completes RELEASE ([#1426](https://github.com/descope/descope-js/issues/1426)) ([a88ca4b](https://github.com/descope/descope-js/commit/a88ca4b0a5c461df4a21b56ba53361622095b7cc)), closes [descope/etc#16353](https://github.com/descope/etc/issues/16353)
+
+## [3.69.2](https://github.com/descope/descope-js/compare/web-component-3.69.1...web-component-3.69.2) (2026-06-22)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.50.6`
+## [3.69.1](https://github.com/descope/descope-js/compare/web-component-3.69.0...web-component-3.69.1) (2026-06-21)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.50.5`
+
+### Bug Fixes
+
+* **web-component:** bump @descope/flow-scripts to 1.0.17 RELEASE ([#1421](https://github.com/descope/descope-js/issues/1421)) ([13b11d3](https://github.com/descope/descope-js/commit/13b11d37953047ad004e88e0bb288bb49fe58323))
+
+## [3.69.0](https://github.com/descope/descope-js/compare/web-component-3.68.0...web-component-3.69.0) (2026-06-18)
+
+
+### Features
+
+* **web-component:** support new doesnt-contains realtime operator RELEASE ([#1417](https://github.com/descope/descope-js/issues/1417)) ([fe37871](https://github.com/descope/descope-js/commit/fe378717bb2bc3492ca1f4c2a1f438b0f647e4be)), closes [descope/backend#1348](https://github.com/descope/backend/issues/1348) [descope/etc#16322](https://github.com/descope/etc/issues/16322) [descope/backend#1348](https://github.com/descope/backend/issues/1348) [#1348](https://github.com/descope/descope-js/issues/1348)
+
 ## [3.68.0](https://github.com/descope/descope-js/compare/web-component-3.67.1...web-component-3.68.0) (2026-06-14)
 
 ### Dependency Updates